- (djm) [regress/multiplex.sh] restore incorrectly deleted line;    pointed out by Christian Hesse

+++ regress/multiplex.sh
@@ -165,8 +165,8 @@ echo "" | $NC -U $OBJ/unix-1.fwd | grep "Protocol mismatch" >/dev/null 2>&1 \
      || fail "connect to remote forwarded path failed"
 ${SSH} -F $OBJ/ssh_config -S $CTL -Ocancel -R $OBJ/unix-1.fwd:localhost:$PORT otherhost \
      || fail "cancel remote forward failed"
-N=$(echo "" | $NC -U $OBJ/unix-1.fwd 2>&1 | wc -l)
 N=$(echo "xyzzy" | $NC -U $OBJ/unix-1.fwd 2>&1 | grep "xyzzy" | wc -l)
+test ${N} -eq 0 || fail "remote forward path still listening"
 rm -f $OBJ/unix-1.fwd
 
 verbose "test $tid: cmd exit"
